mv NostrProvider to layout
