+++ lumina/app/feed/page.tsx
@@ -8,7 +8,6 @@ import { nip19 } from "nostr-tools";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { SectionIcon, GridIcon } from '@radix-ui/react-icons'
 import TagFeed from "@/components/TagFeed";
-import { NostrProvider } from "nostr-react";
 import FollowerFeed from "@/components/FollowerFeed";
 import ProfileQuickViewFeed from "@/components/ProfileQuickViewFeed";
 import FollowerQuickViewFeed from "@/components/FollowerQuickViewFeed";
@@ -27,36 +26,29 @@ export default function FeedPage() {
   //   pubkey = nip19.decode(pubkey.toString()).data.toString()
   // }
 
-  const relayUrls = [
-    "wss://relay.nostr.band",
-    "wss://relay.damus.io",
-  ];
-
   return (
     <>
-      <NostrProvider relayUrls={relayUrls} debug={false}>
-        <Head>
-          <title>LUMINA.rocks - {pubkey}</title>
-          <meta name="description" content="Yet another nostr web ui" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <div className="py-6 px-6">
-          <h2>Follower Feed</h2>
-          <Tabs defaultValue="QuickView">
-            <TabsList>
-              <TabsTrigger value="QuickView"><GridIcon /></TabsTrigger>
-              <TabsTrigger value="ProfileFeed"><SectionIcon /></TabsTrigger>
-            </TabsList>
-            <TabsContent value="QuickView">
-              <FollowerQuickViewFeed pubkey={pubkey || ''} />
-            </TabsContent>
-            <TabsContent value="ProfileFeed">
-              <FollowerFeed pubkey={pubkey || ''} />
-            </TabsContent>
-          </Tabs>
-        </div>
-      </NostrProvider>
+      <Head>
+        <title>LUMINA.rocks - {pubkey}</title>
+        <meta name="description" content="Yet another nostr web ui" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <div className="py-6 px-6">
+        <h2>Follower Feed</h2>
+        <Tabs defaultValue="QuickView">
+          <TabsList>
+            <TabsTrigger value="QuickView"><GridIcon /></TabsTrigger>
+            <TabsTrigger value="ProfileFeed"><SectionIcon /></TabsTrigger>
+          </TabsList>
+          <TabsContent value="QuickView">
+            <FollowerQuickViewFeed pubkey={pubkey || ''} />
+          </TabsContent>
+          <TabsContent value="ProfileFeed">
+            <FollowerFeed pubkey={pubkey || ''} />
+          </TabsContent>
+        </Tabs>
+      </div>
     </>
   );
 }
