+++ lumina/app/onboarding/page.tsx
@@ -1,25 +1,14 @@
 "use client";
 
 import { CreateSecretKeyForm } from "@/components/onboarding/createSecretKeyForm";
-import { Button } from "@/components/ui/button";
-import { NostrProvider } from "nostr-react";
-
 
 export default function OnboardingHome() {
-
-  const relayUrls = [
-    "wss://relay.nostr.band",
-    "wss://relay.damus.io",
-  ];
-
   return (
     <>
-      <NostrProvider relayUrls={relayUrls} debug={false}>
-        <div className="flex flex-col items-center py-6 px-6">
-          <h1>Step 1: Create your secret key</h1>
-          <CreateSecretKeyForm />
-        </div>
-      </NostrProvider>
+      <div className="flex flex-col items-center py-6 px-6">
+        <h1>Step 1: Create your secret key</h1>
+        <CreateSecretKeyForm />
+      </div>
     </>
   );
 }
