+++ lumina/app/search/[searchTag]/page.tsx
@@ -8,7 +8,6 @@ import { nip19 } from "nostr-tools";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { SectionIcon, GridIcon } from '@radix-ui/react-icons'
 import TagFeed from "@/components/TagFeed";
-import { NostrProvider } from "nostr-react";
 import FollowerFeed from "@/components/FollowerFeed";
 import ProfileQuickViewFeed from "@/components/ProfileQuickViewFeed";
 import FollowerQuickViewFeed from "@/components/FollowerQuickViewFeed";
@@ -32,27 +31,20 @@ export default function SearchPage() {
   //   pubkey = nip19.decode(pubkey.toString()).data.toString()
   // }
 
-  const relayUrls = [
-    "wss://relay.nostr.band",
-    "wss://relay.damus.io",
-  ];
-
   return (
     <>
-      <NostrProvider relayUrls={relayUrls} debug={false}>
-        <Head>
-          <title>LUMINA.rocks</title>
-          <meta name="description" content="Yet another nostr web ui" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <div className="py-6 px-6">
-          <div className='grid grid-cols-1 gap-6' >
-            <SearchProfilesBox searchTag={searchTag.toString()} />
-            <SearchNotesBox searchTag={searchTag.toString()} />
-          </div>
+      <Head>
+        <title>LUMINA.rocks</title>
+        <meta name="description" content="Yet another nostr web ui" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <div className="py-6 px-6">
+        <div className='grid grid-cols-1 gap-6' >
+          <SearchProfilesBox searchTag={searchTag.toString()} />
+          <SearchNotesBox searchTag={searchTag.toString()} />
         </div>
-      </NostrProvider>
+      </div>
     </>
   );
 }
